MoreBehaviors
I coded a defense behavior, how to select between behaviors when it knows about the player. I also set up a hitscan system to decide when it can see the player. I just googled it and used one of the official Unity tutorials on raycasting. If it sees the player, it becomes aware of the player and triggers a combat mode (right now, I only have a defense behavior set up). If it hasn't seen the player in a while, it'll forget the player and move into its previous status, either guard or patrol.