Basic Attack
I set up a basic attack behavior. The test Ai simply points at the player and shoots, and if it loses sight of the player it simply charges at the last spot it saw the player. I think. This one was a lot harder to playtest since it actively tries to kill me.