Hearing and Ammo
The Ai can now hear gunfire, and I finally figured out how to fix whatever broke the ammo pickup like a month ago. It gets an index for all weapons in the map with a certain name, such as smg, and gives ammo to whichever one is being held by the player. The fix could never work in a multiplayer game, but this isn't a multiplayer game.