More Gravity Work
I want to have spaceship orbit planet orbit star, however this caused problems because the movement of the planet would significantly damage the orbit of the spacecraft because I'm using 2 body physics. I solved it by correcting for the planet's movement, but that somehow created a seperate problem where the planet stops moving and teleports towards the star when the game starts. If I move the planet around after the game begins, it works like it should.